Separating types and enums into different files
