+++ src/components/Game/GameMode.ts
@@ -1,0 +1,4 @@
+export enum GameMode {
+	Match = 'match',
+	Infinite = 'infinite',
+}
