+++ src/components/Game/GameStatus.ts
@@ -1,0 +1,8 @@
+export enum GameStatus {
+	InitialScreen = 0,
+	SelectOpponentMode = 1,
+	SelectKeys = 2,
+	SelectGameMode = 3,
+	Playing = 4,
+	GameOver = 5,
+}
