+++ src/components/Game/OpponentMode.ts
@@ -1,0 +1,4 @@
+export enum OpponentMode {
+	Machine = 'machine',
+	Player = 'opponent',
+}
