+++ src/components/Game/PlayerKeysType.ts
@@ -1,0 +1,6 @@
+export type PlayerKeys = {
+	upKey: string;
+	downKey: string;
+	leftKey: string;
+	rightKey: string;
+};
